+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add Tailscale
 - Add More Source Packages
 - Resize Root Partisi To 2GB
+- Change MihomoTProxy To Nikki
 
 **Changelog** (07-02-2025)